Review fixes

+++ src/ast_calc.h
@@ -3,7 +3,7 @@
 
 #include "abstract_calc.h"
 
-long double expr_ast_calc (abstract_expr_calc_t* abstract_expr_calc);
+int expr_ast_calc (abstract_expr_calc_t* abstract_expr_calc);
 
 void expr_ast_destroy (abstract_expr_calc_t* abstract_expr_calc);
