+++ src/jit_calc.h
@@ -3,7 +3,7 @@
 
 #include "abstract_calc.h"
 
-long double expr_jit_calc (abstract_expr_calc_t* abstract_expr_calc);
+int expr_jit_calc (abstract_expr_calc_t* abstract_expr_calc);
 
 void expr_jit_destroy (abstract_expr_calc_t* abstract_expr_calc);
